fix: Remove defers in config_changed
Issue #47 describes a problem where Multus tries to restart the pod
after applying NetworkAttachmentDefinitions, but the unit is stuck
for about 15 minutes in a failure loop.

The logs show that `config_changed` is reemited, causing the custom
`nad_config_changed` event to be emited, calling the Multus library.

At that point, Multus is unable to properly read the
NetworkAttachmentDefinitions and fails with an uncaught exception,
crashing the charm. The unit is then stuck in a loop until a timeout
occurs on the Juju controller side.

This change removes all defers that could happen in the `config_changed`
hook. Most defers were not necessary, as `pebble_ready` events would
cover them.

The only defer that was previously critical was in the case of missing
storage. The defer was replaced by listening to the
`config_storage_attached` event.